+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Change where MAPL prefix is added in ./infrastructure/field
 - Change all `MAPL_GridCompAddVarSpec` calls in Gridcomps with `MAPL_GridCompAddSpec` calls to avoid exposting VariableSpec in `use MAPL`. Required changes to interface for `MAPL_GridCompAddSpec` and acg
 - Removed old interfaces MAPL\_RTRN, MAPL\_Vrfy, MAPL\_ASRT, mapl\_ExceptionHandling\_mod
 - Removed old interfaces MAPL_RTRN, MAPL_Vrfy, MAPL_ASRT, mapl_ExceptionHandling_mod
@@ -19,9 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   infrastructure/geom/, infrastructure/field/, infrastructure/field_bundle/, infrastructure/regridder_mgr/,
   enums/, superstructure/state/, superstructure/generic/. Updated 36 files to use mapl_enums_api instead
   of mapl_Enums_internal. All public API exports remain unchanged.
-
-### Changed
-
 - Renamed modules with mapl_<name>Impl_mod to mapl_<name>_mod
 - Eliminitade "shim" interface libraries.
 - Aligned API umbrellas with actual content.  Earlier refactoring had